refactor: ERD 변경에 따른 자료형의 변화로 테스트 코드 수정
MenuRepositoryTest와 MenuServiceTest에서 ERD 변경에 따른 자료형이 불일치하는 부분 수정

+++ src/test/java/com/example/cafe/repository/MenuRepositoryTest.java
@@ -29,7 +29,7 @@ class MenuRepositoryTest {
          */
 
         String name = "아메리카노";
-        Integer price = 5000;
+        Long price = 5000L;
 
         Menu menu = Menu.builder()
                 .name(name)
@@ -61,10 +61,10 @@ class MenuRepositoryTest {
          */
 
         String name = "아메리카노";
-        Integer price = 5000;
+        Long price = 5000L;
 
         Menu menu = Menu.builder()
-                .id(1)
+                .id(1L)
                 .name(name)
                 .price(price)
                 .build();
